% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPLS.R
\name{KPLS}
\alias{KPLS}
\title{Title KPLS IN FINAL}
\usage{
KPLS(x_sam, y_sam, pwav)
}
\arguments{
\item{x_sam}{MATRIX}

\item{y_sam}{VECTOR}

\item{pwav}{a w*av from PLSR1, w_average}
}
\value{
list
}
\description{
Title KPLS IN FINAL
}
