% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gaussian.R
\name{kriging_gaussian}
\alias{kriging_gaussian}
\title{Title a kernel with p=2}
\usage{
kriging_gaussian(x_sample, y_sample)
}
\arguments{
\item{x_sample}{matrix}

\item{y_sample}{vector}
}
\value{
list
}
\description{
Title a kernel with p=2
}
