% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.R
\name{LHD}
\alias{LHD}
\title{LHD}
\usage{
LHD(num, fac, method = c("maxmin", "mincorr"))
}
\arguments{
\item{num}{number of design point}

\item{fac}{k}

\item{method}{maxmin means maximum the minimum distance
mincorr means minimum the column correlation}
}
\value{
latin hyper design
}
\description{
This function is about to build a lhd
}
\details{
This function is about to build a lhd

no
}
