% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR1.r
\name{PLSR1}
\alias{PLSR1}
\title{Title Partial Least Square Regression}
\usage{
PLSR1(x, y)
}
\arguments{
\item{x}{matrix}

\item{y}{vector}
}
\value{
just loadings
}
\description{
Title Partial Least Square Regression
}
