% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_kriging.r
\name{Predictor}
\alias{Predictor}
\title{Title predictor of KPLS AND guassian kriging}
\usage{
Predictor(model, input)
}
\arguments{
\item{model}{a model from KPLS or guassian kriging}

\item{input}{x need to be create y_hat}
}
\value{
the prediction of x
}
\description{
Title predictor of KPLS AND guassian kriging
}
